Extract result into new file
